Partie Raspberry/Ironcar/requirements.txt
tensorflow==1.1.0
keras==2.1.5
Pillow
pandas
numpy
h5py
adafruit-pca9685
evdev